requirements.txt
python-decouple==3.8
httpx==0.27.2
loguru==0.7.2
pyyaml==6.0.2
pydantic==2.9.2
